separated product tests into component, view, and updater
fixed breaking test in firefox
