+++ test/fixtures/product-fixture.js
@@ -24,19 +24,19 @@ const testProduct = {
     {
       name: 'Print',
       values: [
-        { value: 'sloth' },
-        { value: 'shark' },
-        { value: 'cat' },
-      ]
+        {value: 'sloth'},
+        {value: 'shark'},
+        {value: 'cat'},
+      ],
     },
     {
       name: 'Size',
       selected: 'small',
       values: [
-        { value: 'small' },
-        { value: 'large' }
-      ]
-    }
+        {value: 'small'},
+        {value: 'large'},
+      ],
+    },
   ],
   variants: [
     {
@@ -52,13 +52,13 @@ const testProduct = {
       selectedOptions: [
         {
           name: 'Print',
-          value: 'sloth'
+          value: 'sloth',
         },
         {
           name: 'Size',
-          value: 'small'
-        }
-      ]
+          value: 'small',
+        },
+      ],
     },
     {
       id: 'Z2lkOi8vc2hvcGlmeS9Qcm9kdWN0VmFyaWFudC8xMjM0Ng==',
@@ -73,13 +73,13 @@ const testProduct = {
       selectedOptions: [
         {
           name: 'Print',
-          value: 'shark'
+          value: 'shark',
         },
         {
           name: 'Size',
-          value: 'small'
-        }
-      ]
+          value: 'small',
+        },
+      ],
     },
     {
       id: 'Z2lkOi8vc2hvcGlmeS9Qcm9kdWN0VmFyaWFudC8xMjM0Nw==',
@@ -94,13 +94,13 @@ const testProduct = {
       selectedOptions: [
         {
           name: 'Print',
-          value: 'shark'
+          value: 'shark',
         },
         {
           name: 'Size',
-          value: 'large'
-        }
-      ]
+          value: 'large',
+        },
+      ],
     },
     {
       id: 'Z2lkOi8vc2hvcGlmeS9Qcm9kdWN0VmFyaWFudC8xMjM0OA==',
@@ -111,15 +111,15 @@ const testProduct = {
       selectedOptions: [
         {
           name: 'Print',
-          value: 'cat'
+          value: 'cat',
         },
         {
           name: 'Size',
-          value: 'small'
-        }
-      ]
+          value: 'small',
+        },
+      ],
     },
-  ]
-}
+  ],
+};
 
 export default testProduct;
